// frontend/src/app/model/Family.ts
export class Family {
  name: string;
  dob: string;
  address: string;
  googleLocation: string;
  nativeAddress: string;
  mobileNumber: number;
  qualification: string;
  profession: string;
  organization: string;
  professionNature: string;
  houseCategory: string;
  incomeRange: string;
  familyId: string;
  memberId: string;
  email: string;
  area: string;
  memberDate: string;
  inBangaloreSince: string;
  memberStatus: string;
  designation: string;

}